Add systemd handling of RT throttling
Add an option to indicate that stalld was started by
a systemd service file and that the service file is
responsible for turning off RT throttling and restoring
it after stopping stalld.

The -S/--systemd option tells stalld not to fool with
RT throttling, that's handled for us by the systemd start/stop
logic. If the --systemd flag is not used, turn off RT throttling
within stalld and try to restore with the on_exit() handler (this
is known to not work when killing stalld with -9 signal).

Signed-off-by: Clark Williams <williams@redhat.com>