stalld: Add the -g/--granularity option
Originally, stalld was sleeping for one second between each check for
starving threads. This was good because the threshold value is in the
order of seconds. However, on larger systems (100+ CPUs), the time to
parse the collect and parse the sched_debug was causing the tool to
use barelly 100% of the CPU time.

The one-second granularity, however, is not necessary for all cases.
For example, one user may allow a five seconds granularity to save CPU
time.

With a value different from one, the granularity might affect the
perceived threshold value. For example, on a system with a threshold
of 30 and granularity of five, the perceived threshold will be at most
35 seconds.

Signed-off-by: Daniel Bristot de Oliveira <bristot@kernel.org>