stalld: Implement the single-threaded mode
This mode serves for power saving because it is single threaded.
The drawback of this mode, when compared with the conservative mode,
is that it reduces the detection's granularity for the boost duration
(not a problem if the granularity is higher than the boost duration,
though). It will also boost a single thread at each loop, instead of
all tasks that might be starving for more than the threshold. This
behavior is not bad, though! As multiple boosts in the same loop might
as well delay the detection of other starving threads.

Signed-off-by: Daniel Bristot de Oliveira <bristot@kernel.org>