utils: Fix bounds check on cpu and end_cpu variables
A buffer overflow exists due to a missing bounds check
on both cpu and end_cpu.

Running:
stalld -f -c 0-nr_cpus

is enough to trigger the overflow.

The two variables serve as an index and so are limited to values
between 0 and nr_cpus - 1. Reject values outside of this range.

Signed-off-by: Fernando Pacheco <fpacheco@redhat.com>