stalld: Add the overload control to the single-threaded mode
Currently, the single-threaded mode boosts a single thread per cpu at
a time. This is a disadvantage when compared to the multithreaded
methods. To overcome this limitation, this patch adds an overload
control.

When more than one starving thread is detected, the cpu is marked
as overloaded. After boosting the longest starving threads at the
main thread loop, check if any of the CPUs were overloaded. If so,
instead of sleeping until the next programmed check, skip the sleep
and check for starving threads again. Thus giving a chance for the
other thread to be selected. While this will consume a little bit
more CPU when the granularity is considerably larger than the boost
period, the chances are that other CPUs will also have new starving
threads.

Signed-off-by: Daniel Bristot de Oliveira <bristot@redhat.com>