stalld: Add the adaptive mode option
The current default mode of stalld uses a single thread when nothing
is happening, but dispatches a per-cpu thread after a starving thread
is waiting for half of the config_starving_threshold.

It does not have a config option because it is the default.

In preparation to make the single-threaded the default mode, add
a config option to select the current default mode, named adaptive
mode.

Signed-off-by: Daniel Bristot de Oliveira <bristot@redhat.com>