stalld: Fix log message on boost_cpu_starving_vector()
boost_cpu_starving_vector() is printing the following messages:

  stalld: boosting cpu 0: pid: 0 starving for 1623936064
  stalld: boosting cpu 1: pid: 180105 starving for 15
  stalld: boosting cpu 2: pid: 0 starving for 1623936064
  stalld: boosting cpu 3: pid: 0 starving for 1623936064
  stalld: boosting cpu 4: pid: 0 starving for 1623936064
  stalld: boosting cpu 5: pid: 0 starving for 1623936064
  stalld: boosting cpu 6: pid: 0 starving for 1623936064
  stalld: boosting cpu 7: pid: 0 starving for 1623936064

But they are wrong in two ways. Firstly, the message is
not about boosting, but about starving threads. Secondly,
the pid == 0 means that no thread is starving on that CPU.

Fix it by not printing if pid == 0, and adjust the message
to say that the task is starving instead of boosting. It
is also using a better indentation as well.

With this patch in place, the message now looks like:

  stalld: 	 cpu 1: pid: 180105 starving for 15

No functional changes.

Signed-off-by: Daniel Bristot de Oliveira <bristot@redhat.com>