stalld v1.12
Most of the changes are cleanups. The most visible change is that
the stalld's systemd service now starts the daemon with FIFO:10
priority. This change is necessary to avoid having stalld starving or
being indirectly blocked on a real-time mutex owned by a starving
thread. The latter behavior indeed happened in the multi-threaded mode.

Changes:
  stalld.8: fix diff cruft left in manpage source
  stalld.c: clean up handling of nr_running
  stalld.c: remove duplicate parameter to fill_waiting_task()
  stalld: Add error handling in get_cpu_idle_time()
  stalld.service: Run stalld as sched_fifo via systemd
  packaging: clean up Makefiles and rpm specfile
  stalld: Always print current function for info messages
  stalld: Always print current function for warn messages
  stalld: Always print current function for die messages
  utils: change PATHMAX to 4096

Signed-off-by: Daniel Bristot de Oliveira <bristot@redhat.com>