utils.c: Fail if user is not root
Stalld needs access to some files and calls some system calls that
require root permission. Currently, stalld does not check if it is
running as root, and so stalld fails with a message error on some
permission error.

To make it clear to the user that stalld needs superuser permission,
now it fails to start if the effective user is not root, printing a
message about it.

Signed-off-by: Daniel Bristot de Oliveira <bristot@redhat.com>