stalld v1.13
This is version v1.13 of stalld. It includes some significant changes.

The service script now restarts stalld if it fails.

The single-threaded mode was improved with an overload control.
Previous, the single-threaded mode boosts a single thread per cpu
at a time. This was a disadvantage when compared to the
multithreaded methods. Now, when more than one starving thread is
detected, the cpu is marked as overloaded. After boosting the
longest starving threads at the main thread loop, check if any of
the CPUs were overloaded. If so, instead of sleeping until the next
programmed check, skip the sleep and check for starving threads
again. Thus giving a chance for the other thread to be selected.
While this will consume a little bit more CPU when the granularity
is considerably larger than the boost period, the chances
are that other CPUs will also have new starving threads.

With the overload control in place, the single-threaded mode
becomes the best one and is now the default.

The kernel v5.13 has changed the path to /proc/sched_debug to
$debugfs/sched/debug. This version is aware of this change and
tries to access both /proc/sched_debug, and $debugfs/sched/debug.

Changes:
  01eb3ce utils.c: Make the path to sched_debug path dynamic
  6d0b834 utils.c: Fail if user is not root
  4316bc8 stalld: Make single-threaded mode the default one
  47723ac stalld: Fallback to the adaptive mode if force_fifo is selected
  c04eea8 stalld: Use the last mode set in the cmdline
  398c5a9 stalld: Add the adaptive mode option
  8f49a3a stalld: Add the overload control to the single-threaded mode
  d9e10a6 stalld: Fix log message on boost_cpu_starving_vector()
  1e1732f stalld.service: Always restart stalld on exit

Signed-off-by: Daniel Bristot de Oliveira <bristot@redhat.com>