stalld: Add an option to easily set stalld as SCHED_DEADLINE
Add the -R/--reservation option. This option requires one argument,
an integer between 10 and 90, which is the percentage of CPU time
that will be allocated to the stalld daemon itself via SCHED_DEADLINE.

The dl_runtime and dl_period of the stalld daemon will be computed
by the daemon itself to achieve the -R "percentage" of CPU time at
at every "granularity." So the user does not need to think about
the specific values, while avoiding known problems, such as
setting the period higher than 4 seconds.

Signed-off-by: Daniel Bristot de Oliveira <bristot@kernel.org>