utils.c: Exit early if enabling HRTICK fails when using SCHED_DEADLINE
stalld allows the boost runtime that can be specified to be at a
maximum of 1ms. At this granularity, the HZ tick will not be enough to
make scheduling decisions at the needed time. Therefore HRTICK needs
to be enabled to ensure that SCHED_DEADLINE tasks only run for a duration
of "dl->runtime" and not end up overrunning their allowed bandwidth.

Without HRTICK, the priority boosted task will potentially
end up overrunning it's runtime bandwidth. stalld
cannot reliably use SCHED_DEADLINE without enabling HRTICK.

If HRTICK cannot be enabled for some reason, we should not be allowing
stalld to run in a mode where it is doing priority boosting using
SCHED_DEADLINE. Add the necessary code for this

We also need to enable HRTICK only when stalld is using SCHED_DEADLINE.
If stalld is using SCHED_FIFO for priority boosting, it mimics SCHED_DEADLINE
using calls to nanosleep. The nanosleep system call sets up its own hrtimers
independent of HRTICK.

Signed-off-by: Sharan Turlapati <sturlapati@vmware.com>