stalld: Add queue track eBPF based backend
Instead of reading the sched/debug file, load an eBPF program
that tracks the enqueue/dequeue of tasks in the sched fair
and rt schedulers.

The information is stored in BPF_MAP_TYPE_ARRAY map, with the CPU as key.

Q: Why not a per-cpu map?
A: Because tasks can be enqueued in a remote CPU.

The only thing done in eBPF is getting the minimal necessary information
while hooked to the kernel events, while keeping all the
detection/boosting logic in user-space. This choice was made to avoid
overhead in the CPU queueing tasks, keeping it as much as possible
in user-space, in a house-keeping CPU.

My preliminary tests using osnoise did not notice a change in the
osnoise metric in an isolated CPU.

This patch only adds the feature, it cannot be enabled yet, so:

No Functional changes are expected.

Signed-off-by: Daniel Bristot de Oliveira <bristot@kernel.org>