stalld: v1.18
This is the version 1.18 of stalld.

The main feature of this version is the addition of the
queue_track backend (and the backend abstraction as well).

Instead of parsing sched/debug, which is problematic on very
large systems, the queue_track backend monitors the queue/dequeue
of threads in the [per-cpu] runqueue using eBPF to do so.

The main advantage is to reduce the cpu time used by stalld.
The drawback is that part of stalld processing is now on the
monitored CPUs. Nevertheless, the overhead is < 1 us, as measured
via osnoise. Thus, it is a valid option for those that
want to avoid osnoise consuming much CPU.

The queue_track backend is disabled by default. To enable it,
use the -b Q option. It can eventually become the default option
after some versions of in-field testing.

For those distros that do not support eBPF, the stalld version
1.17 will be maintained as a long term release including fixes.

- c100d63 queue_track: Use LIBBPF_MAJOR/MINOR_VERSION to detect deprecated functions
- d15f141 utils: Close file descriptor
- da88d5e stalld: Fix function name of daemonize()
- e07a26b docs: Fix typo in the manual
- 0641ce5 queue_track: Use bpf_map__resize on older libbpf versions
- 89c290c utils: Let tgid to arrive at the fill proccess comm
- 772730d stalld: Fix log message on idle detection
- 2cdd114 stalld: Add -b/--backend option
- 53db11b stalld: Add queue track eBPF based backend
- a95940e stalld: Add fill_process_comm helper
- 91ceb4d stalld: Include regex.h on stalld.h
- 94bbc55 stalld: Get nr of cpus only once
- 959a4a4 stalld: Add the backend abstraction
- f5c345b sched_debug: Move sched debug functions to an specific file

Signed-off-by: Daniel Bristot de Oliveira <bristot@kernel.org>