stalld: v1.19.1
This version improves the makefile, mainly to properly
support the eBPF-based backend.

It is now possible to compile stalld with or without eBPF,
and a logic to automatically set according to te eBPF support
for the architecture.

It is possible to manually set your preference via USE_BPF flag
at compilation time. E.g.:

 # make USE_BPF=[0,1]

Changes:

- 0a228ea Conditionalize BPF and queue_track build per architecture
- bcc157c clean up Makefile install logic and add .bz2 to .gitignore

It has no features, so a minor relese.

Signed-off-by: Daniel Bristot de Oliveira <bristot@kernel.org>