src/utils.c: fix off-by-one error in buffer allocation
The buffer size allocation calculation in find_debugs_sched_debug()
fails to account for an extra '/' used when joining path fragments,
resulting in allocating a buffer one byte too small. This error was
not found locally because the Makefile uses FORTIFY_SOURCE=2, while
the package build uses FORTIFY_SOURCE=3.

Change the size calculation to account for the extra '/' and for
additional paranoia, change to using snprintf().

Signed-off-by: Clark Williams <williams@redhat.com>