Makefile:  refactor CFLAGS/LDFLAGS for local and koji builds
The original makefile had CFLAGS and LDFLAGS that duplicated the
ones used for distro build (hardened) but it also has the defines
that enabled the bpf logic (USE_BPF). The dist-git builds completely
over-rode CFLAGS and LDFLAGS and the builds lost the defines.

Change the specfile to pass CFLAGS as RPMCFLAGS and LDFLAGS as
RPMLDFLAGS and use some conditional logic in the makefile to set
CFLAGS/LDFLAGS appropriately.

Signed-off-by: Clark Williams <williams@redhat.com>