bpf: Improve task tracking in sched_switch
The previous implementation in the sched_switch handler was unable to
track tasks that were already running when stalld was started. It only
updated the context switch count for tasks already in the tracking queue,
leading to an incomplete view of the system's active tasks.

This patch introduces a new helper function, update_or_add_task(), to
provide more robust task lifecycle management directly within the
sched_switch handler.

This function now dynamically adds any running task to the tracking queue
if it's not already present. As a result of this more accurate state
tracking, the function also removes tasks from the queue if they are found
to be in a non-running state, which prevents the accumulation of stale
entries.

By calling this function for both the prev and next tasks in
sched_switch, the BPF program now maintains a more accurate and
dynamic view of the tasks on the run queue, correctly tracking both
newly scheduled tasks and those that were active before stalld began
monitoring.

Resolves: RHEL-96213

Signed-off-by: Wander Lairson Costa <wander@redhat.com>
Signed-off-by: Clark Williams <williams@redhat.com>