Refactor: Remove duplicated find_debugfs function
The find_debugfs() function was a wrapper that cached the debugfs
mount point. Its functionality was duplicated by
find_debugfs_mount_point(), which also located the debugfs path.

To eliminate this redundancy and simplify the code, this commit
replaces calls to find_debugfs() with find_debugfs_mount_point()
and removes the now-unused find_debugfs() function.

A forward declaration for find_debugfs_mount_point() was added to
maintain the build order.

Signed-off-by: Wander Lairson Costa <wander@redhat.com>