bpf: Reuse enqueue_task() in update_or_add_task()
Refactor update_or_add_task() to call the existing enqueue_task()
function when a new task needs to be added to the tracking queue.

This change eliminates duplicated code, making the logic for adding a
new task consistent and centralized within enqueue_task(). The local
variables in update_or_add_task() were also simplified as part of
this cleanup.

Signed-off-by: Wander Lairson Costa <wander@redhat.com>