bpf: a task is dequeued just by setting its pid to 0
The user reads the pid to know that there is no data in the queue,
so there is no need to set the other fields, prio and ctxswc, so
clearing of `prio` and `ctxswc` is redundant, as these
fields are not read when the pid is 0. The memory barrier is also not
needed.

This commit removes the unnecessary code to simplify the dequeuing
logic.

Signed-off-by: Wander Lairson Costa <wander@redhat.com>