stalld: Remove forced log-only mode with DL-server
When a DL-server is detected, stalld currently logs a message and
forces itself into log-only mode. This behavior is too restrictive
and prevents stalld from performing its throttling duties in environments
where a DL-server is present but throttling is still desired.

This patch removes the automatic switch to log-only mode. Now, stalld
will only print a notification message upon detecting a DL-server,
but will otherwise continue to function with its configured settings.
The decision to run in log-only mode is left to the user's
configuration.

Resolves: RHEL-105604

Signed-off-by: Wander Lairson Costa <wander@redhat.com>