Makefile: Remove redundant GCC version check
The minimum GCC version check was introduced to handle the `-fcf-protection`
flag, which was not available in older compiler versions. However, a more
robust method that directly tests for compiler support for this flag has
been implemented.

This commit removes the now-redundant GCC version check, simplifying the
Makefile and relying on the more accurate feature detection.

Signed-off-by: Wander Lairson Costa <wander@redhat.com>
Cc: Derek Barbosa <debarbos@redhat.com>
Cc: John Kacur <jkacur@redhat.com>
Cc: Juri Lelli <juri.lelli@redhat.com>
Cc: Chunsheng Luo <luochunsheng@ustc.edu>
Signed-off-by: Clark Williams <williams@redhat.com>